xtensa: fix TLB miss handling
This change fixes the TLB miss handling within the exception.

As currently implemented, the 2 levels page table translation works
as long as virtual addresses are less than 0x400000, because it
yields PTEVADDR which are within the top level page mapped in one of
the wired TLB way. For addresses greater than or equal to 0x400000,
it yields PTEVADDR for which there is not a top level page mapped.

For example, with CONFIG_XTENSA_MMU_PTEVADDR == 0x20000000 and
l1_page_table address at 0x0000c000, we will map the top level as
"vpn" == 0x20080000, "ppn" == 0x0000c000
If an address needing translation is 0x400000, it yields an
exception with PTEVADDR == 0x20001000, and within the exception
handler, the TLB auto-refill will attempt to translate 0x20080004
for which there would not be a fixed mapping.
WRONG: mapping exists

Signed-off-by: William Tambe <williamt@cadence.com>